Create tagToLanguage.py
This program was designed to take possible language codes and create a lookup table connecting the code, with the language name, the code's components and any Wikidata language IDs assigned to each code. Currently it creates a list of dictionaries with everything above except the IDs. The lookup table is not optimum at this point. (also just found out langcodes python library can probably do a better job as some codes are deprecated)